Rename Populator to StreamsTableWrapper
